[FLINK-21851][table-runtime] Refactor BinaryRowDataKeySelector in testing
This closes #15256
