+++ flink-table/flink-table-runtime-blink/src/test/java/org/apache/flink/table/runtime/operators/window/WindowOperatorContractTest.java
@@ -28,14 +28,15 @@ import org.apache.flink.table.runtime.generated.NamespaceAggsHandleFunction;
 import org.apache.flink.table.runtime.generated.NamespaceAggsHandleFunctionBase;
 import org.apache.flink.table.runtime.generated.NamespaceTableAggsHandleFunction;
 import org.apache.flink.table.runtime.generated.RecordEqualiser;
+import org.apache.flink.table.runtime.keyselector.RowDataKeySelector;
 import org.apache.flink.table.runtime.operators.window.assigners.MergingWindowAssigner;
 import org.apache.flink.table.runtime.operators.window.assigners.WindowAssigner;
 import org.apache.flink.table.runtime.operators.window.triggers.Trigger;
-import org.apache.flink.table.runtime.util.BinaryRowDataKeySelector;
 import org.apache.flink.table.types.logical.BigIntType;
 import org.apache.flink.table.types.logical.IntType;
 import org.apache.flink.table.types.logical.LogicalType;
 import org.apache.flink.table.types.logical.VarCharType;
+import org.apache.flink.table.utils.HandwrittenSelectorUtil;
 
 import org.junit.Rule;
 import org.junit.Test;
@@ -194,8 +195,8 @@ public class WindowOperatorContractTest {
 
         LogicalType[] inputTypes =
                 new LogicalType[] {new VarCharType(VarCharType.MAX_LENGTH), new IntType()};
-        BinaryRowDataKeySelector keySelector =
-                new BinaryRowDataKeySelector(new int[] {0}, inputTypes);
+        RowDataKeySelector keySelector =
+                HandwrittenSelectorUtil.getRowDataSelector(new int[] {0}, inputTypes);
         TypeInformation<RowData> keyType = keySelector.getProducedType();
         LogicalType[] accTypes = new LogicalType[] {new BigIntType(), new BigIntType()};
         LogicalType[] windowTypes = new LogicalType[] {new BigIntType(), new BigIntType()};
